add Jest, fix downgrade issue
