+++ jest.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  transformIgnorePatterns: ['<rootDir>/node_modules/*']
+};
